fix: edit datasource for External saas plugins (#38672)
