+++ app/client/src/ce/constants/ReduxActionConstants.tsx
@@ -857,6 +857,8 @@ const DatasourceEditorActionTypes = {
   CREATE_DATASOURCE_INIT: "CREATE_DATASOURCE_INIT",
   CREATE_DATASOURCE_SUCCESS: "CREATE_DATASOURCE_SUCCESS",
   CREATE_DATASOURCE_FROM_FORM_INIT: "CREATE_DATASOURCE_FROM_FORM_INIT",
+  CREATE_DATASOURCE_FROM_FORM_TOGGLE_LOADING:
+    "CREATE_DATASOURCE_FROM_FORM_TOGGLE_LOADING",
   CREATE_TEMP_DATASOURCE_FROM_FORM_SUCCESS:
     "CREATE_TEMP_DATASOURCE_FROM_FORM_SUCCESS",
   UPDATE_DATASOURCE_INIT: "UPDATE_DATASOURCE_INIT",
