+++ app/client/src/ce/pages/Editor/DataSourceEditor/ExternalSaasConnection.tsx
@@ -1,0 +1,3 @@
+export default function ExternalSaasConnection() {
+  return null;
+}
