+++ app/client/src/ee/pages/Editor/DataSourceEditor/ExternalSaasConnection.tsx
@@ -1,0 +1,3 @@
+export * from "ce/pages/Editor/DataSourceEditor/ExternalSaasConnection";
+import { default as CE_ExternalSaasConnection } from "ce/pages/Editor/DataSourceEditor/ExternalSaasConnection";
+export default CE_ExternalSaasConnection;
