+++ app/client/src/pages/Editor/IntegrationEditor/index.css
@@ -1,0 +1,3 @@
+#paragon-connect-frame {
+  pointer-events: auto;
+}
