+++ app/client/src/reducers/entityReducers/datasourceReducer.ts
@@ -126,6 +126,12 @@ const datasourceReducer = createReducer(initialState, {
   ) => {
     return { ...state, loading: true };
   },
+  [ReduxActionTypes.CREATE_DATASOURCE_FROM_FORM_TOGGLE_LOADING]: (
+    state: DatasourceDataState,
+    action: ReduxAction<{ loading?: boolean }>,
+  ) => {
+    return { ...state, loading: !!action.payload.loading };
+  },
   [ReduxActionTypes.UPDATE_DATASOURCE_INIT]: (state: DatasourceDataState) => {
     return { ...state, loading: true };
   },
